MAINT: delete basis module
