--- matgl/basis/__init__.py
@@ -1,3 +1,0 @@
-"""Implementations of math basis functions."""
-
-from ..layers._basis import GaussianExpansion, SphericalBesselFunction, SphericalHarmonicsFunction
